refactor(infra): use stable coreos image "slug"
Instead of updating to the latest stable version, this now points to
whatever image is the latest stable version. This prevents updating the
VM whenever a new stable version is released, but also allows for the
latest stable version to be used whenever the VM is recreated for some
other purpose.